cleaned code and set-up variables for remote server

+++ gestioneForm.php
@@ -10,33 +10,41 @@ $connection = new Connection();
 // TODO verificare errore in rimozione
 if(isset($_POST['nomeRemoveP'])){
 	$full_qualifier= $_POST['nomeRemoveP'];
-	$split = explode('-',$full_qualifier);
-	$table='';
-	switch ($split[0]){
-		case '0':
-			$table='auto';
-			break;
-		case '1':
-			$table='bare';
-			break;
-		case '2':
-			$table='urne';
-			break;
-		case '3':
-			$table='composizioni';
-			break;
-		case '4':
-			$table='cerimonie';
-			break;
-	}
-	$id=$split[1];
-	$del_res=$connection->Query("
-		DELETE FROM $table WHERE id='$id'
-	");
-	$del_error ="Rimozione effettuata con successo";
-	if(!$del_res){
-		$del_error="Errore di rimozione";
-	}
+	$del_error='<div class="linea" id="errors"><a href="#nomeRemoveP">';
+	if($full_qualifier== '---'){
+	    $del_error .= 'Selezionare un prodotto da rimuovere';
+    }
+	else{
+        $split = explode('-',$full_qualifier);
+        $table='';
+        switch ($split[0]){
+            case '0':
+                $table='auto';
+                break;
+            case '1':
+                $table='bare';
+                break;
+            case '2':
+                $table='urne';
+                break;
+            case '3':
+                $table='composizioni';
+                break;
+            case '4':
+                $table='cerimonie';
+                break;
+        }
+        $table = $connection->escape($table);
+        $id=$connection->escape($split[1]);
+        $del_res=$connection->Query("
+		    DELETE FROM $table WHERE id='$id'
+	    ");
+        if(!$del_res)
+            $del_error .= 'Errore di rimozione';
+        else
+            $del_error .= 'Rimozione effettuata con successo';
+    }
+	$del_error.= '</a></div>';
 }
 require_once 'bin/Connection.php';
 $connection = new Connection();
@@ -67,6 +75,4 @@ fclose($management_file);
 $man_content = str_replace('<elementlist/>',$option_list,$man_content);
 $man_content = str_replace('<deleteerror/>',isset($del_error)
 	?$del_error:'',$man_content);
-$man_content = str_replace('<adderror/>',isset($add_error)
-	?$add_error:'',$man_content);
 echo $man_content;
